feat(cache): 添加商品热度指数获取功能并优化缓存策略

+++ product-module/src/main/java/com/zhuxi/product/module/interfaces/vo/ProductDetailVO.java
@@ -49,7 +49,7 @@ public class ProductDetailVO {
     @Schema(description = "商品浏览次数")
     private Long viewCount;
     @Schema(description = "商品热度指数")
-    private Long hostScore;
+    private BigDecimal hostScore;
     @Schema(description = "商品图片列表")
     private List<ProductImage> productImages;
